poll loop added, refactored affected hpp and cpp files, start refactor Server.cpp/hpp
