+++ include/Config.hpp
@@ -1,13 +1,23 @@
 #pragma once
 
-// class Client;
-
 #include <sstream>
 #include <fstream>
 #include <vector>
 #include <iostream>
 #include "../include/Location.hpp"
 #include "EventPoll.hpp"
+#include "../include/Server.hpp"
+#include "../include/Location.hpp"
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <netinet/in.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <arpa/inet.h>
+#include <netdb.h>
+#include <poll.h>
 
 class Server;
 
@@ -22,7 +32,7 @@ class Config {
         Config(const Config& other) = default;
         Config& operator=(const Config& other) = default;
         ~Config();
-        int check_config(const std::string &config_file);
+        int checkConfig(const std::string &config_file);
 		std::vector<std::string> tokenize(const std::string &line);
 		bool isFileEmpty(const std::string& fileName);
 		std::vector<Server> parse_config(std::ifstream &file);
