+++ include/EventPoll.hpp
@@ -4,17 +4,23 @@
 #include <poll.h>
 #include <unistd.h>
 
+/**
+ * @brief Structure to specify a file descriptor and event type to be removed.
+ */
 typedef struct s_pollfdToRemove {
 	int fd;
 	int eventType;
 }	t_pollfdToRemove;
 
+/**
+ * @brief A class to manage poll events and their associated file descriptors.
+ */
 class EventPoll {
 
 	private:
-		std::vector<struct pollfd> 			_pollfds;
-		std::vector<t_pollfdToRemove> 		_pollfdsToRemoveQueue;
-		std::vector<pollfd>					_pollfdsToAddQueue;
+		std::vector<struct pollfd> 			_pollfds;						// List of active poll file descriptors
+		std::vector<t_pollfdToRemove> 		_pollfdsToRemoveQueue;			// Queue of file descriptors to be removed.
+		std::vector<pollfd>					_pollfdsToAddQueue;				// Queue of file descriptors to be added.
 
 	public:
 		EventPoll();
