+++ include/utils.hpp
@@ -7,5 +7,6 @@
 # define WRITE 1
 # define WRITE_SIZE 10
 # define READ_SIZE 10
+# define BACKLOG 10
 
 std::string getStatusMessage(int statusCode);
