+++ src/HttpRequest.cpp
@@ -1,11 +1,5 @@
 #include "../include/HttpRequest.hpp"
 
-// HttpRequest::HttpRequest(std::string method, std::string path, std::string version){
-// 	// this->method = method;
-// 	// this->path = path;
-// 	// this->version = version;
-// }
-
 /**
  * @brief       Default constructor for the HttpRequest class.
  * 
@@ -13,52 +7,10 @@
  */
 HttpRequest::HttpRequest(){
 	_strReceived = "";
-	// std::cout << "HttpRequest() " << _strReceived << std::endl;
 }
 
-/**
- * @brief       Destructor for the HttpRequest class.
- * 
- * @details     Ensures proper cleanup of resources used by the HttpRequest object.
- *              Currently, no specific cleanup is needed.
- */
 HttpRequest::~HttpRequest(){}
 
-/**
- * @brief       Assignment operator for the HttpRequest class.
- * 
- * @param       copy        The HttpRequest object to copy from.
- * 
- * @return      A reference to the current HttpRequest object.
- * 
- * @details     Performs a deep copy of the _strReceived and _request fields.
- */
-HttpRequest& HttpRequest::operator=(const HttpRequest& copy){
-	// std::cout << "HttpRequest::operator= 1"  << std::endl;
-
-    if (this != &copy) { // CORRECT?
-	this->_strReceived = copy._strReceived;
-	this->_request = copy._request;
-    }
-	// std::cout << "HttpRequest::operator= 2"  << std::endl;
-    return *this;
-}
-
-/**
- * @brief       Copy constructor for the HttpRequest class.
- * 
- * @param       copy        The HttpRequest object to copy from.
- * 
- * @details     Creates a new HttpRequest object by copying the _strReceived and 
- *              _request fields from the given object.
- */
-HttpRequest::HttpRequest(const HttpRequest& copy){
-	// std::cout << "HttpRequest(const HttpRequest& copy) 1"  << std::endl;
-	this->_strReceived = copy._strReceived;
-	this->_request = copy._request;
-	// std::cout << "PATH :" << this->_request["path"] << std::endl;
-	// std::cout << "HttpRequest(const HttpRequest& copy) 2"  << std::endl;
-}
 
 /**
  * @brief       Sets a key-value pair in the internal request field map.
@@ -78,11 +30,9 @@ void HttpRequest::setField(std::string key, std::string value){
  * @return      The value associated with the key, or an empty string if the key does not exist.
  */
 std::string HttpRequest::getField(std::string key){
-	// std::cout << "KEY IS: " << key << std::endl;
 	auto it = _request.find(key);
 	if (it != _request.end()){
-		// std::cout << "VALUE IS: " << it->second << std::endl;
-		return trim(it->second); //added trim
+		return trim(it->second);
 	}
 	return "";
 }
@@ -135,24 +85,22 @@ void HttpRequest::setHeaderReceived(bool received){
 void HttpRequest::readRequest(std::string request){
 
 	std::istringstream request_stream(request);
-	std::string method, path, version, host; //?
+	std::string method, path, version, host;
 	std::string line;
 	request_stream >> method >> path >> version; 
 	setField("method", method);
 	setField("path", path);
-	setField("version", version);
-	// std::cout << "This is the test line path" << path << std::endl; 
+	setField("version", version); 
 	while(std::getline(request_stream, line)){
 		if (line.empty())
 			break;
-		const unsigned long colon = line.find(':');//changed from int to const unsigned long because flags
+		const unsigned long colon = line.find(':');
 		if (colon != std::string::npos){
 			std::string key = line.substr(0, colon);
 			key = trim(key);
 			std::string value = line.substr(colon + 1);
 			value = trim(value);
 			setField(key, value);
-			// std::cout << "key is: " << key << "       value is: " << value << std::endl;
 		}
 	}
 }
@@ -171,27 +119,6 @@ std::string HttpRequest::trim(std::string& str) {
     return str.substr(first, (last - first + 1));
 }
 
-// int HttpRequest::checkErrors() {
-//     // std::cout << "Checking for errors. Method: |" << getField("method") << "|"<< " Version: " << getField("version") << std::endl;
-//     if (getField("method") != "GET" && method != "POST" && method != "DELETE") {
-//         std::cerr << "Error: Invalid method." << std::endl;
-//         return 405;
-//     }
-//     if (getField("version").empty()) {
-//         std::cerr << "Error: Version is empty or invalid!" << std::endl;
-//         return 400;
-//     }
-//     if (getField("version") != "HTTP/1.1") { //should we accept HTTP/1.0?
-// /*
-// For persistent connections (such as in HTTP/1.1), you would leave the client in the pfds list to handle further requests.
-// For non-persistent connections (such as in HTTP/1.0), it's appropriate to remove the client after processing the request.
-// */
-//         std::cerr << "Error: Invalid HTTP version." << std::endl;
-//         return 400;
-//     }
-//     return 200;
-// }
-
 /**
  * @brief       Determines the content length from an HTTP request string.
  * 
@@ -202,13 +129,19 @@ std::string HttpRequest::trim(std::string& str) {
 int HttpRequest::findContentLength(std::string request){
 	this->readRequest(request);
 	std::string len = this->getField("Content-length");
-	// std::cout << "Content length is " << len << std::endl;
 	if (len != "")
 		return(std::stoi(len));
 	else
 		return 0;
 }
 
+/**
+ * @brief       Sets the string representing the raw received HTTP request.
+ * 
+ * @param       input       The string to set as the raw received HTTP request.
+ * 
+ * @details     Replaces the current value of _strReceived with the input string.
+ */
 void HttpRequest::setStrReceived(std::string input) {
 	_strReceived = input;
 }
