+++ src/Server.cpp
@@ -1,33 +1,28 @@
 #include "../include/Server.hpp"
-#include "../include/Client.hpp"
-#include "../include/HttpRequest.hpp"
-#include <iostream>     // for std::cout
-#include <vector>      // for std::vector
-#include <poll.h>      // for poll() and struct pollfd
-#include <sys/types.h> // for basic types
-#include <sys/socket.h> // for socket-related functions and structures
-#include <netinet/in.h> // for sockaddr_in, sockaddr_storage
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <arpa/inet.h>
-#include <netdb.h>
 
-#define BACKLOG 10
-
-Server::Server(){
-}
+Server::Server(){}
 
 Server::~Server(){}
 
-/*Function to create and return a listener socket
-- status: Result of getaddrinfo(), checks if address info is valid
-- hints: addrinfo structure with settings for creating socket
-- servinfo: Linked list of address structures
-- newConnect: Pointer to iterate through servinfo to create a socket
-- serverSocket: File descriptor for server socket
-- opt: Option value for setsockopt() to allow address reuse*/
+/**
+ * @brief Initializes and returns a listener socket for the server.
+ *
+ * This function sets up a TCP listener socket using the server's configured port.
+ * It configures address information, creates the socket, sets socket options, 
+ * binds the socket to the specified address, and begins listening for incoming
+ * connections. The function returns the file descriptor for the created listener
+ * socket on success, or exits the program if binding fails for all addresses.
+ * 
+ * status: Result of getaddrinfo(), checks if address info is valid
+ * hints: addrinfo structure with settings for creating socket
+ * servinfo: Linked list of address structures
+ * newConnect: Pointer to iterate through servinfo to create a socket
+ * serverSocket: File descriptor for server socket
+ * opt: Option value for setsockopt() to allow address reuse
+ *
+ * @return The file descriptor for the created listener socket on success, 
+ *         or -1 if an error occurs during listening setup.
+ */
 int Server::get_listener_socket(){
 	int status;
 	struct addrinfo hints;
@@ -66,18 +61,32 @@ int Server::get_listener_socket(){
 	return serverSocket;
 }
 
-/*Function to remove a file descriptor from the poll vector
-- pfds: Vector of pollfd structures
-- i: Index of the file descriptor to remove from the poll vector*/
+
+/**
+ * @brief Removes a pollfd element from the given vector at the given index i.
+ *
+ * This function swaps the element to be deleted with the last element in the vector,
+ * and then pops the last element to remove it. This is done to avoid having to shift
+ * all the elements after the one to be deleted.
+ *
+ * @param pfds The vector of pollfd elements to modify
+ * @param i The index of the element to remove
+ * @todo this will be replaced by the pollEvent class function and deleted
+ */
 void Server::del_from_pfds(std::vector<struct pollfd> &pfds, int i){
 	pfds[i] = pfds.back(); // Move the last one to the deleted spot
 	pfds.pop_back(); // Remove the last one (which is now duplicated)
 }
 
-/*Function to report readiness of the server and set up the listener socket
-- pfds: Vector of pollfd structures where listener will be added
-- listener: The file descriptor for the listener socket
-TODO do we need to treturn listeren or should listener be a private variable?*/
+/**
+ * @brief Reports that the server is ready to accept connections
+ *
+ * This function sets up a listening socket and adds it to the EventPoll
+ * so that it can be monitored for incoming connections.
+ *
+ * @param eventPoll The EventPoll to add the listening socket to
+ * @return The file descriptor of the listening socket
+ */
 int Server::reportReady(EventPoll &eventPoll){
 	int listener = get_listener_socket(); // Set up and get a listening socket
 	if (listener == -1){
@@ -85,7 +94,7 @@ int Server::reportReady(EventPoll &eventPoll){
 	}
     // Add the listener to EventPoll
     eventPoll.addPollFdEventQueue(listener, POLLIN);
-    return listener;
+    return listener; //TODO do we need to treturn listeren or should listener be a private variable?
 }
 
 /**
@@ -123,6 +132,7 @@ void Server::handleNewConnection(EventPoll &eventPoll){
  * @param eventPoll The EventPoll object containing the pollfds
  * @param i The index into the pollfds vector
  * @todo  implement client->closeConnection(); in client
+ * @todo  divide into smaller functions doing one thing
  */
 void Server::handlePollEvent(EventPoll &eventPoll, int i) {
     Client *client = nullptr;
@@ -189,6 +199,15 @@ void Server::handlePollEvent(EventPoll &eventPoll, int i) {
 }
 
 /**
+ * @brief Checks and updates the server's root based on the given path.
+ *
+ * This function compares a given path with the server's index. If they
+ * match, the function returns immediately. Otherwise, it iterates over
+ * the server's locations to find a matching path. If a matching location
+ * is found and it is not empty, the server's root is updated to the root
+ * defined in the location.
+ *
+ * @param path The path to check against the server's locations.
  * @todo figure out when to reset server information to default
  */
 void Server::checkLocations(std::string path) {
@@ -204,4 +223,254 @@ void Server::checkLocations(std::string path) {
 			}
 		}
 	}
+}
+
+/**
+ * @brief Process an HTTP request received from a client.
+ *
+ * This function processes an HTTP request string and takes the appropriate
+ * action based on the request method and path. It returns an HTTP status code
+ * indicating the result of the request processing.
+ *
+ * If the request method is not allowed for the given path, a 405 Method Not
+ * Allowed response is sent to the client. If the request path does not match
+ * any of the server's locations, a 404 Not Found response is sent to the
+ * client. If there is an error parsing the request, a 400 Bad Request response
+ * is sent to the client.
+ *
+ * @param client The client object associated with the request.
+ * @param request The HTTP request string received from the client.
+ * @param HttpRequest The HttpRequest object associated with the client.
+ * @return The HTTP status code indicating the result of the request processing.
+ */
+int Server::processClientRequest(Client &client, const std::string& request, HttpRequest* HttpRequest) {
+	std::istringstream requestStream(request);
+	std::cout <<" This is request "<< request << std::endl;
+	std::string method, path, version;
+	HttpResponse response;
+	requestStream >> method >> path >> version;
+	HttpRequest->readRequest(request);
+	checkLocations(path);
+	int status = validateRequest(method, version);
+	if (status != 200) {
+		sendFileResponse(client.getSocket(), "www/html/500.html", status);  //change to a config ones?
+		return status;
+	}
+	if (method == "GET" && std::find(this->_allowed_methods.begin(), this->_allowed_methods.end(), "GET") != this->_allowed_methods.end())
+		//check with this endpoint am I allowed to use get?
+		return handleGetRequest(client, path, HttpRequest); //?? what locations should be passed
+	if (method == "POST" && std::find(this->_allowed_methods.begin(), this->_allowed_methods.end(), "POST") != this->_allowed_methods.end())
+	//check with this endpoint am I allowed to use post?
+		return handlePostRequest(client, path, HttpRequest);
+	if (method == "DELETE" && std::find(this->_allowed_methods.begin(), this->_allowed_methods.end(), "DELETE") != this->_allowed_methods.end())
+	//check with this endpoint am I allowed to use delete?
+		return handleDeleteRequest(client, path, HttpRequest);
+	response.buildResponse();
+	// check poll if I can write?
+	return 0;
+}
+
+/**
+ * @brief Handles an HTTP GET request from a client.
+ *
+ * This function takes a request path and determines how to handle the GET request.
+ * If the path is a directory, the function sends the contents of the directory
+ * as the response body. If the path is a file, the function reads the contents
+ * of the file and sends it as the response body. If the path does not exist, a
+ * 404 Not Found response is sent with a simple HTML page indicating that the
+ * file was not found.
+ *
+ * @param client The client object associated with the request.
+ * @param path The path of the GET request.
+ * @param request The HttpRequest object associated with the client.
+ * @return The HTTP status code indicating the result of the request processing.
+ * @todo writing needs to go through the poll loop not working yet
+ */
+int Server::handleGetRequest(Client &client, const std::string& path, HttpRequest* request) {
+	
+	std::string filepath = this->getRoot() + '/' + path;
+	if (path == "/") {
+		filepath = this->getRoot() + '/' + this->getIndex();
+	}
+	if (path.rfind("/cgi-bin/", 0) == 0) { //change to config
+		client.startCgi(request);
+		return 0;
+	}
+	std::ifstream file(filepath);
+	if (!file) {
+		std::cerr << "Error: File not found for path " << filepath << std::endl;
+		sendFileResponse(client.getSocket(), "www/html/404.html", 404);
+		return 404;
+	}
+	// setFileResponse()
+	sendFileResponse(client.getSocket(), filepath, 200);//needs to be set respionse because need to go back to poll loop
+	return 200;
+}
+
+/**
+ * @brief Send a file response to the client.
+ *
+ * This function reads the contents of the specified file and sends it as the body
+ * of an HTTP response to the client. If the file does not exist, a 404 Not Found
+ * response is sent with a simple HTML page indicating that the file was not found.
+ *
+ * @param clientSocket The socket to send the response to.
+ * @param filepath The path to the file to send.
+ * @param statusCode The HTTP status code to send in the response.
+ * @todo check sendHeaders() for passing content type not hard coded
+ */
+void Server::sendFileResponse(int clientSocket, const std::string& filepath, int statusCode) {
+	std::string fileContent = readFileContent(filepath);
+	if (fileContent.empty()) {
+		sendHeaders(clientSocket, 404, "text/html");
+		sendBody(clientSocket, "<html><body>404 - File Not Found</body></html>");
+	} else {
+		sendHeaders(clientSocket, statusCode, "text/html");
+		sendBody(clientSocket, fileContent);
+	}
+	close(clientSocket);
+}
+
+/**
+ * @brief Read the contents of a file into a string.
+ *
+ * This function reads the contents of a file specified by the given filepath
+ * and returns the contents as a string. If the file does not exist, an error
+ * message is printed to stderr and an empty string is returned.
+ *
+ * @param filepath The path to the file to read.
+ * @return The contents of the file as a string.
+ */
+std::string Server::readFileContent(const std::string& filepath) {
+    std::ifstream file(filepath);
+    if (!file) {
+        std::cerr << "Error: File not found: " << filepath << std::endl;
+        return "";
+    }
+    std::ostringstream buffer;
+    buffer << file.rdbuf(); //read file by bytes, go back to poll, check if finished reading
+	//request->_readyToSendBack = true;
+	std::cout.flush();
+    return buffer.str();
+}
+
+/**
+ * @brief Send HTTP headers to the client.
+ *
+ * This function sends the HTTP headers (status line and content type) to
+ * the client. The status line is constructed with the provided status code
+ * and the corresponding status message. The "Content-Type" header is set to
+ * the provided content type. If no content type is provided, it defaults
+ * to "text/html".
+ *
+ * @param clientSocket The socket to send the headers to.
+ * @param statusCode The HTTP status code to send.
+ * @param contentType The content type to send (optional, defaults to "text/html").
+ */
+void Server::sendHeaders(int clientSocket, int statusCode, const std::string& contentType = "text/html") {
+    std::string statusMessage = getStatusMessage(statusCode);
+    std::ostringstream headers;
+    headers << "HTTP/1.1 " << statusCode << " " << statusMessage << "\r\n";
+    // For persistent connections (such as in HTTP/1.1), you would leave the client in the pfds list to handle further requests.
+    // For non-persistent connections (such as in HTTP/1.0), it's appropriate to remove the client after processing the request.
+    headers << "Content-Type: " << contentType << "\r\n\r\n"; // Default is text/html
+    std::string headersStr = headers.str();
+    send(clientSocket, headersStr.c_str(), headersStr.size(), 0);
+}
+
+/**
+ * @brief Send the body of the HTTP response to the client.
+ *
+ * This function sends the contents of the provided string as the body
+ * of the HTTP response to the client.
+ *
+ * @param clientSocket The socket to send the body to.
+ * @param body The body content to send.
+ */
+void Server::sendBody(int clientSocket, const std::string& body) {
+    send(clientSocket, body.c_str(), body.size(), 0);
+}
+
+/**
+ * @brief Validates the HTTP method and version of the request.
+ * 
+ * This function checks if the provided HTTP method is one of the
+ * supported methods ("GET", "POST", "DELETE") and if the version
+ * is "HTTP/1.1". If the method is not supported, it returns a 405
+ * status code indicating "Method Not Allowed". If the version is
+ * not "HTTP/1.1", it returns a 400 status code for "Bad Request".
+ * On successful validation, it returns a 200 status code for "OK".
+ * 
+ * @param method The HTTP method to validate.
+ * @param version The HTTP version to validate.
+ * @return The HTTP status code indicating the result of the validation.
+ */
+int Server::validateRequest(const std::string& method, const std::string& version) {
+	if (method != "GET" && method != "POST" && method != "DELETE") {
+		std::cerr << "Error: Unsupported HTTP method." << std::endl;
+		return 405; // Method Not Allowed
+	}
+
+	if (version != "HTTP/1.1") {
+		std::cerr << "Error: Invalid HTTP version. Only HTTP/1.1 is supported." << std::endl;
+		return 400;
+	}
+	return 200;
+}
+
+/*
+Extract the path from the request (usually the file or resource to be deleted).
+Check if the requested resource exists. If not, return a 404 Not Found response.
+For security reasons, it's essential to implement some form of authorization to ensure that only authorized users can delete resources.
+If the user isn't authorized, return a 401 Unauthorized or 403 Forbidden.
+Ensure the resource being deleted is a valid file, directory, or resource that can be deleted. For example, check if it’s a file in a specific directory on the server.
+Optionally, check if the file can be deleted (i.e., it’s not locked or in use).
+Perform the deletion. For files or directories, use appropriate system calls to delete the resource.
+If it’s a database entry or another type of resource, ensure the record is properly deleted from the data store.
+Check if there are any issues while deleting (e.g., file permissions, file not found, or resource is locked).
+If an error occurs during the deletion process, return a 500 Internal Server Error or a more specific status code.
+If the deletion was successful, send a 200 OK or 204 No Content response.
+If there was an issue, return a corresponding error code:
+403 Forbidden: If the user is not allowed to delete the resource.
+404 Not Found: If the file or resource does not exist.
+500 Internal Server Error: If there was an error during the deletion process.
+It’s often useful to log the deletion operation for auditing purposes, especially if your server manages important data.
+As with the POST request, decide whether to close the connection or keep it alive based on the HTTP version or the Connection header.
+*/
+int Server::handleDeleteRequest(Client &client, const std::string& path, HttpRequest* Http) {
+	(void)client;
+	(void)path;
+	(void)Http;
+
+	return 0;
+}
+
+/*Extract the HTTP headers to understand what type of data is being sent (e.g., Content-Type, Content-Length).
+If the body size exceeds a predefined limit (for example, from a config file), return an error like 413 Request Too Large.
+Ensure that the Content-Length header is present and valid. This will tell you how much data to expect in the body.
+Read the body of the request from the socket. You might need to handle partial reads (i.e., the body could arrive in chunks).
+Form Data: If Content-Type is application/x-www-form-urlencoded, parse the form fields and their values.
+JSON Data: If the request contains application/json, you can parse the JSON data and extract the necessary information.
+File Uploads: If the Content-Type is multipart/form-data, handle file uploads (you'll need to parse the file boundaries and save the file to disk).
+Form Submission: Save data to a database, perform an action, or return a response to confirm the form was submitted.
+File Upload: Save the file to a specific directory and generate a success/failure response.
+Once processing is complete, send an appropriate response to the client. 
+Success (200 OK): If the request was processed successfully.
+Error (400 Bad Request): If there was a problem with the data.
+Error (500 Internal Server Error): If something went wrong on the server side.
+Decide whether to close the connection or keep it alive (based on HTTP version or a Connection header).
+*/
+int Server::handlePostRequest(Client &client, const std::string& path, HttpRequest* Http) {
+	
+	(void)client;
+	(void)path;
+	(void)Http;
+
+	// if (!Http->findContentLength(Http->_strReceived) || Http->findContentLength(Http->_strReceived) <= 0) //Can be equal to 0?
+	// 	return 400;
+	// // if (Http->findContentLength(Http->_strReceived) > this->getMaxlength()) //add after parsing
+	// // 	return 413;
+	// if (Http->getField("Content-type") == "application/x-www-form-urlencoded")
+
+	return 0;
 }
